Status bar expanded: Fix background opacity/transparency:
Background transparency depends on the transparency state of notification_panel_bg,
this means, the current transparency is choosed transparency + state of notification_panel_bg,
also it isn`t possible to set a transparency less than the transparency of notification_panel_bg,
even if transparency set to 0% it isn`t fully opaque.

So the Idea is to add additional drawables with value ff000000, (fully opaque, black),
if transparency is set to 0%, the background is fully opaque now,
and the the choosed transparency depends on a fully opaque state,
which will fix the wrong transparency value, too.

If "Theme default" is enabled, the original notification_panel_bg drawable will be loaded from system/theme,
to be compatible to Theme Chooser Themes.

Change-Id: Ie4ee061ca5f4805550397ad543ee67f09967afa1