Battery icon: Change charging bolt and status text appearience while charging:
The charging animation shows if the device is currently charging,
on the battery normal icon I prefer to know the current battery status, too,
instead of replacing the status text by the charging bolt during charge.

So always show the status text on the normal icon,
except charging animation is disabled.