Lock screen and status bar: Notification and status area color improvements and fixes, (1/2):
Lock screen:
- Colorize the notification more icon

Status bar:
- Colorize the notification more icon
- Colorize notification and status icons:
  - Add a possibility to only colorize when the icon is an greyscale icon,
    (now the options are: Never, when greyscale or always)
- Fix colorizing options:
  Sometimes the icons didn`t respected the user choice,
  e.g. a notification icon got colorized even if colorizing notification icons was disabled

I tested this with different notifications and status icons,
but I don`t be sure if this will work with all notifications and status icons.
So far everything I tested worked as expected, but I don`t be satisfied with the expectation itself,
related to the app which "creates" the notification, the status bar notification icon could be greyscale, white or colored,
colorizing an colored icon could be looking very bad, that`s the main reason for the colorizing options,
unfortunately I didn`t found a possibility to check if the icon is a notification or a status icon,
status icons are nearly always, (always ? ) white, so status icons should be colorized always.
As a workaround for now, the status icons are respecting the colorizing options, too.
I`ll take a look at it later.
Even if a check for the status bar more icon isn`t needed, and there is no problem to colorize it always,
I decided it should also respects the colorizing options as long as I don`t find a possibility to check for the status icons,
so it`s behavior is the same as the behavior of the status icons.