Power menu customizations, (1/2):
Included:
- Button arrangement, (up to 8 buttons)
- Background color
- Icon normal color
- Icon enabled/selected color
- Ripple color
- Text color
- Don`t execute the action automatically,
  if the power menu contains one button only

This will also colorize the background of the reboot conformation- and shutdown dialogs,
but no other colors of these dialogs, as the text color of these dialogs is
white, (title, message and single choice items), changing the background color
to white will result in invisible texts.
I`ll also (try) to colorize these dialogs, too,
but this will need some additions inside the (Alert/Progress/...)Dialog classes,
this should be done in a seperate commit.