Status bar: Battery customizations, (1/2):
- Battery icon:
  - Battery portrait
  - Battery landscape
  - Circle
    - Dotted circle:
      - Interval, (interval = 0 will show the circle undotted)
      - Length
  - Hidden
- Show/hide battery text
- Enable/disable cut out battery text
- Enable/disable battery charging animation
- Additional battery bar:
  - Show/hide the battery bar
  - Show/hide the battery bar when on lock screen

This also takes effect of the status bar when on the lock screen,
but no effect on the battery of the expanded header.
The SystenUI Tuner setting ("Show embedded battery percentage")
takes only effect on the battery of the expanded header.

The battery level text of the status bar when on the lock screen
will only be shown, if the battery text is hidden, battery is shown and
the battery is charging.

The battery icon and battery bar will respect the global status bar colors,
(normal status bar and when on lock screen)
The battery level text of the status bar when on the lock screen,
uses the global status bar text color, while the embedded text and charging indicator
uses the global status bar battery text color.

The Battery icon additions are based on CyanogenMod.