Quick settings: Custom colors, (1/2):
Custom colors:
- Background
- Accent
- Text
- Icon
- Ripple effect

Changes:
- Dnd Detail view:
  - Zen mode button text colors:
    - Normal state: Fully opaque color
    - Selected state: Accent color
- Brightness slider:
  - Remove the mirror

The brightness slider thumb uses a VectorDrawable
with 2 different colorized paths,
one uses the QS panel background color,
while the other is used for drawing the thumb icon itself.
As far as I know it isn`t possible to set a color to a
specific path programmatically, so when colorizing the thumb
both paths will be using the same color, which makes the
thumb looking very bad, removing the path for the background
isn`t a solution, because the brightness progress bar will be
visible behind the thumb. So I splitted the thumb icon into
2 seperate VectorDrawables, which are wrapped into a LayerDrawable,
so in this case it`s possible to set a different color for both
VectorDrawables.